解封 批量封禁

+++ api/src/main/java/com/luckygames/wmxz/gamemaster/dao/handler/ActionTypeHandler.java
@@ -30,7 +30,7 @@ public class ActionTypeHandler implements TypeHandler<ActionType> {
 
     @Override
     public void setParameter(PreparedStatement ps, int i, ActionType param, JdbcType jdbcType) throws SQLException {
-        ps.setInt(i, (byte) param.code());
+        ps.setInt(i, param.code());
     }
 
 }
